% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H4PackagesOfoegbuKingsley-package.R
\docType{package}
\name{H4PackagesOfoegbuKingsley-package}
\alias{H4PackagesOfoegbuKingsley}
\alias{H4PackagesOfoegbuKingsley-package}
\title{H4PackagesOfoegbuKingsley: Creating a Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kingsley Ofoegbu \email{kd6889a@student.american.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
