% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_function.R
\name{my_seq_plot}
\alias{my_seq_plot}
\title{Plot Function}
\usage{
my_seq_plot(x)
}
\arguments{
\item{x}{Dataframe}
}
\value{

}
\description{
This function plots the sequence of the data frame comes out as a line plot
}
