% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sequence.R
\name{myseq_n}
\alias{myseq_n}
\title{Sequence Function}
\usage{
myseq_n(x, n)
}
\arguments{
\item{x}{Integers}

\item{n}{Length}
}
\value{

}
\description{
This function takes account a vector input containing the first 3 numeric
of this sequence and a positive (>0) integer n denoting the final nth element of the sequence to calculate
}
